+++ demisto_sdk/commands/pre_commit/hooks/ruff.py
@@ -2,7 +2,7 @@ from copy import deepcopy
 from pathlib import Path
 from typing import Any, Dict, Set
 
-from demisto_sdk.commands.pre_commit.hooks.hook import Hook, join_files
+from demisto_sdk.commands.pre_commit.hooks.hook import NIGHTLY_MODE, Hook, join_files
 
 
 class RuffHook(Hook):
@@ -38,7 +38,7 @@ class RuffHook(Hook):
             hook["args"] = [
                 f"--target-version={self._python_version_to_ruff(python_version)}",
             ]
-            if self.all_files:
+            if self.mode == NIGHTLY_MODE:
                 hook["args"].append("--config=nightly_ruff.toml")
             else:
                 hook["args"].append("--fix")